IFD production defects fixed.

+++ TYR_2.0_18.5/TY_03_REPORTS.py
@@ -2623,12 +2623,12 @@ class PlotCanvas(FigureCanvas):
                          ax.set_xlim(0,int((float(x[0]))))
                          ax.set_ylim(0,int((float(x[1])*9.81)))                     
                      elif(self.unit_type == "MPA"):
-                         ax.set_xlim(0,float(x[0]))
-                         ax.set_ylim(0,float((float(x[1])*9.81))) 
+                         ax.set_xlim(0,int(x[0]))
+                         ax.set_ylim(0,int((float(x[1])*9.81))) 
                      else:
                          if(self.test_type=="COF"):
-                             ax.set_xlim(0,float(int(x[0])))
-                             ax.set_ylim(0,float(x[1]))
+                             ax.set_xlim(0,int(int(x[0])))
+                             ax.set_ylim(0,int(x[1]))
                          else:
                              ax.set_xlim(0,float(float(x[0])*0.1))
                              ax.set_ylim(0,float(x[1]))
